requirements.txt
Flask==2.3.2
gunicorn==21.2.0
# Required to avoid build failure
setuptools==69.5.1
wheel==0.43.0
pip==25.1.1

# Core dependencies
numpy==1.26.4
pillow==10.0.1

# OCR & Image Processing
opencv-python-headless==4.8.1.78
easyocr==1.7.1

# MySQL
mysql-connector-python==8.0.33

# Utility
python-dotenv==1.0.1
